Simplify corrections module
